# perf: speed up adding fields, reduce memalloc if field name is already prefixed with "app."