--- a/beeline_test.go
+++ b/beeline_test.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/honeycombio/beeline-go/trace"
 	"github.com/honeycombio/libhoney-go/transmission"
 
 	libhoney "github.com/honeycombio/libhoney-go"
@@ -107,11 +106,6 @@ func BenchmarkBeelineAddField(b *testing.B) {
 
 	ctx, _ := StartSpan(context.Background(), "parent")
 
-	b.Run("oldAddField/whatever", func(b *testing.B) {
-		for n := 0; n < b.N; n++ {
-			oldAddField(ctx, "foo", 1)
-		}
-	})
 	b.Run("AddField/no-prefix", func(b *testing.B) {
 		for n := 0; n < b.N; n++ {
 			AddField(ctx, "foo", 1)
@@ -131,22 +125,6 @@ func BenchmarkBeelineAddField(b *testing.B) {
 			AddField(ctx, "app.foo", 1)
 		}
 	})
-}
-
-func oldAddField(ctx context.Context, key string, val interface{}) {
-	span := trace.GetSpanFromContext(ctx)
-	if span != nil {
-		if val != nil {
-			namespacedKey := "app." + key
-			if valErr, ok := val.(error); ok {
-				// treat errors specially because it's a pain to have to
-				// remember to stringify them
-				span.AddField(namespacedKey, valErr.Error())
-			} else {
-				span.AddField(namespacedKey, val)
-			}
-		}
-	}
 }
 
 func setupLibhoney(t testing.TB) *transmission.MockSender {